// src/app/services/api/model/get-user-dto.ts
/**
 * EfCoreLibraryAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetUserDto { 
    id?: number;
    name?: string | null;
    firstName?: string | null;
    email?: string | null;
    birthday?: string | null;
}